Goodbye .env
