--- agixt/.env.example
@@ -1,62 +1,0 @@
-# =========================
-# WORKING DIRECTORY
-# =========================
-WORKING_DIRECTORY=./WORKSPACE
-WORKING_DIRECTORY_RESTRICTED=True
-
-# =========================
-# EXTENSIONS: HUGGINGFACE
-# =========================
-HUGGINGFACE_API_KEY=
-HUGGINGFACE_AUDIO_TO_TEXT_MODEL=facebook/wav2vec2-large-960h-lv60-self
-
-# =========================
-# EXTENSIONS: SELENIUM
-# =========================
-SELENIUM_WEB_BROWSER=chrome
-
-# =========================
-# EXTENSIONS: TWITTER
-# =========================
-TW_CONSUMER_KEY=my-twitter-consumer-key
-TW_CONSUMER_SECRET=my-twitter-consumer-secret
-TW_ACCESS_TOKEN=my-twitter-access-token
-TW_ACCESS_TOKEN_SECRET=my-twitter-access-token-secret
-
-# =========================
-# EXTENSIONS: GITHUB
-# =========================
-GITHUB_API_KEY=
-GITHUB_USERNAME=
-
-# =========================
-# EXTENSIONS: SENDGRID
-# =========================
-SENDGRID_API_KEY=
-SENDGRID_EMAIL=
-
-# =========================
-# EXTENSIONS: MICROSOFT 365
-# =========================
-MICROSOFT_365_CLIENT_ID=
-MICROSOFT_365_CLIENT_SECRET=
-MICROSOFT_365_REDIRECT_URI=
-
-# =========================
-# EXTENSIONS: SEARXNG
-# =========================
-SEARXNG_INSTANCE_URL=https://searx.work
-
-# =========================
-# VOICE SETTINGS
-# =========================
-
-# BrianTTS
-USE_BRIAN_TTS=True
-
-# Mac OS
-USE_MAC_OS_TTS=False
-
-# ElevenLabs
-ELEVENLABS_API_KEY=
-ELEVENLABS_VOICE=Josh
