+++ agixt/__init__.py
@@ -2,7 +2,7 @@ from .AGiXT import AGiXT
 from .Memories import Memories
 from .Chain import Chain
 from .Commands import Commands
-from .Config.Agent import Agent
+from .Agent import Agent
 from .Config import Config
 from .Embedding import Embedding
 from .CustomPrompt import CustomPrompt
